Cast to int/double for n_ratings fixed.

+++ app/src/main/java/it/polito/madcourse/group06/models/userprofile/UserProfile.kt
@@ -53,7 +53,7 @@ data class UserProfile(
 
     @get:PropertyName("n_ratings")
     @set:PropertyName("n_ratings")
-    var n_ratings: Int,
+    var n_ratings: Double,
 
     @get:PropertyName("comments_services_rx")
     @set:PropertyName("comments_services_rx")
